+++ includes/ray.h
@@ -6,7 +6,7 @@
 /*   By: daspring <daspring@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/15 15:16:05 by daspring          #+#    #+#             */
-/*   Updated: 2024/11/26 13:44:22 by daspring         ###   ########.fr       */
+/*   Updated: 2024/11/27 13:51:01 by daspring         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,6 +28,6 @@ typedef struct s_ray
 
 // t_ray	create_ray(t_tuple E, t_tuple P);
 t_ray	create_ray(size_t x_pixel, size_t y_pixel);
-t_tuple	ray_at_t(t_ray *ray, float t);
+t_tuple	ray_at_t(t_ray *ray, double t);
 
 #endif
