+++ libft/ft_atof_mod.c
@@ -6,7 +6,7 @@
 /*   By: daspring <daspring@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/24 23:25:33 by daspring          #+#    #+#             */
-/*   Updated: 2024/11/15 17:13:34 by daspring         ###   ########.fr       */
+/*   Updated: 2024/11/27 13:51:01 by daspring         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,13 +14,13 @@
 // #include "../include/fractol.h"
 
 static int		find_signum(char *s, int *signum);
-static float	get_number_part_greater_one(char *s, int *pos, float number);
-static float	get_number_part_smaller_one(char *s, int *pos, float number);
+static double	get_number_part_greater_one(char *s, int *pos, double number);
+static double	get_number_part_smaller_one(char *s, int *pos, double number);
 static int		is_whitespace(char c);
 
-float	ft_atof_mod(char *s, int *error)
+double	ft_atof_mod(char *s, int *error)
 {
-	float	number;
+	double	number;
 	int		signum;
 	int		pos;
 
@@ -58,7 +58,7 @@ static int	find_signum(char *s, int *signum)
 	return (pos);
 }
 
-static float	get_number_part_greater_one(char *s, int *pos, float number)
+static double	get_number_part_greater_one(char *s, int *pos, double number)
 {
 	while (s[*pos] && ft_isdigit(s[*pos]))
 	{
@@ -70,9 +70,9 @@ static float	get_number_part_greater_one(char *s, int *pos, float number)
 	return (number);
 }
 
-static float	get_number_part_smaller_one(char *s, int *pos, float number)
+static double	get_number_part_smaller_one(char *s, int *pos, double number)
 {
-	float	decimal_place;
+	double	decimal_place;
 
 	decimal_place = 0.1;
 	while (s[*pos] && ft_isdigit(s[*pos]))
