+++ sources/canvas/lighting.c
@@ -6,7 +6,7 @@
 /*   By: daspring <daspring@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/31 12:17:20 by llacsivy          #+#    #+#             */
-/*   Updated: 2024/11/26 21:14:00 by daspring         ###   ########.fr       */
+/*   Updated: 2024/11/27 13:51:01 by daspring         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,7 +39,7 @@ t_color	calc_diffuse_color(t_hit_obj *hit_obj, t_object *light)
 	t_color	diffuse_color;
 	t_color	effective_color;
 	t_tuple	light_vec;
-	float	light_dot_normal;
+	double	light_dot_normal;
 	t_tuple	temp;
 
 	effective_color = color_mult(hit_obj->obj->color, light->s_light.intensity);
@@ -57,10 +57,10 @@ t_color	calc_diffuse_color(t_hit_obj *hit_obj, t_object *light)
 t_color	calc_specular_color(t_hit_obj *hit_obj, t_object *light, t_ray *ray)
 {
 	t_tuple	light_vec;
-	float	light_dot_normal;
+	double	light_dot_normal;
 	t_tuple	dir_pt_light;
 	t_tuple	reflect_vec;
-	float	reflect_dot_eye;
+	double	reflect_dot_eye;
 	t_tuple	pt;
 	t_tuple	temp;
 	t_tuple	temp2;
@@ -101,7 +101,7 @@ t_tuple	calc_reflect_vec(t_tuple *incomming, t_tuple *normal_vec)
 {
 	t_tuple	reflect_vec;
 	t_tuple	subtrahend;
-	float	ray_dot_normal;
+	double	ray_dot_normal;
 
 	ray_dot_normal = tuple_dot(incomming, normal_vec);
 	subtrahend = tuple_scale2(2 * ray_dot_normal, normal_vec);
