+++ sources/maths/tuple_1.c
@@ -6,7 +6,7 @@
 /*   By: daspring <daspring@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: Invalid date        by                   #+#    #+#             */
-/*   Updated: 2024/11/26 20:33:38 by daspring         ###   ########.fr       */
+/*   Updated: 2024/11/27 13:51:01 by daspring         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -84,7 +84,7 @@ t_tuple	tuple_subtr2(t_tuple *minuend, t_tuple *subtrahend)
 	return (c);
 }
 
-t_tuple	*tuple_scale(float k, t_tuple *a)
+t_tuple	*tuple_scale(double k, t_tuple *a)
 {
 	t_tuple	*c;
 
@@ -96,7 +96,7 @@ t_tuple	*tuple_scale(float k, t_tuple *a)
 	return (c);
 }
 
-t_tuple	tuple_scale2(float k, t_tuple *a)
+t_tuple	tuple_scale2(double k, t_tuple *a)
 {
 	t_tuple	c;
 
@@ -107,7 +107,7 @@ t_tuple	tuple_scale2(float k, t_tuple *a)
 	return (c);
 }
 
-t_tuple	set_tuple(float x, float y, float z, int w)
+t_tuple	set_tuple(double x, double y, double z, int w)
 {
 	t_tuple	tuple;
