+++ sources/maths/tuple_2.c
@@ -6,7 +6,7 @@
 /*   By: daspring <daspring@student.42heilbronn.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/10 19:53:23 by llacsivy          #+#    #+#             */
-/*   Updated: 2024/11/26 19:23:25 by daspring         ###   ########.fr       */
+/*   Updated: 2024/11/27 13:51:01 by daspring         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 #include "../../includes/maths.h"
 #include "../../includes/miniRT.h"
 
-float	tuple_magni(t_tuple *a)
+double	tuple_magni(t_tuple *a)
 {
 	return (sqrt(a->x * a->x + a->y * a->y + a->z * a->z));
 }
@@ -52,7 +52,7 @@ bool	is_normalized(t_tuple *vec)
 		return (false);
 }
 
-float	distance(t_tuple *a, t_tuple *b)
+double	distance(t_tuple *a, t_tuple *b)
 {
 	t_tuple	temp;
